% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_js.R
\name{put_text_at_3rd_column_}
\alias{put_text_at_3rd_column_}
\title{Put text column at third position of values dataframe}
\usage{
put_text_at_3rd_column_(lst)
}
\arguments{
\item{lst}{A list with dataframe element \strong{values}, used in the internal of \code{\link{qr_parse_js_}}.}
}
\value{
A list with dataframe element \strong{values}.
}
\description{
This function is only called by \code{\link{qr_parse_js_}}, to put the text
column at the third position of the values dataframe.
}
\keyword{internal}
