% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_js_src_path_}
\alias{qr_js_src_path_}
\title{Get the included JS source for \code{jsQR}.}
\usage{
qr_js_src_path_()
}
\value{
The path to the included version of the \code{jsQR} library.
}
\description{
Get the included JS source for \code{jsQR}.
}
\keyword{internal}
