% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_js_src_update}
\alias{qr_js_src_update}
\title{Update the JS source for \code{jsQR} from the upstream repo.}
\usage{
qr_js_src_update(local_path = getwd(), reset = F)
}
\arguments{
\item{local_path}{The working folder to save the new version of the \code{jsQR} library.}

\item{reset}{Reverts to the included version if problems arise.}
}
\value{
The path to a new version of the jsQR library.
}
\description{
Changes the \code{\link{options}} setting to a new version of the JS source code,
as determined from the canonical repository on GitHub. If you wish to modify
the JS source yourself or download from another URL, you can instead call
\code{options(list("quadrangle.js_src" = your_local_file_path))}
}
