% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_js.R
\name{qr_parse_corners_}
\alias{qr_parse_corners_}
\title{(Internal) Crop and resize large images to QR corner neighborhoods}
\usage{
qr_parse_corners_(mgk, code_pts)
}
\arguments{
\item{mgk}{A \pkg{magick} image object.}

\item{code_pts}{A dataframe of columns \strong{x} and \strong{y}, with each row identifying a QR pattern corner, usually generated by \code{qr_scan_cpp_magick}.}
}
\value{
A list with the altered image, the \strong{x}/\strong{y} anchor pixels, and a resizing flag
}
\description{
This is an internal function used by \code{\link{qr_scan_js_from_corners}}.
}
\keyword{internal}
