% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qr_pb_}
\alias{qr_pb_}
\title{(Internal) Progress bar wrapper}
\usage{
qr_pb_(prefix, n)
}
\arguments{
\item{prefix}{Which scanning engine is running?}

\item{n}{How many combinations of arguments will be tried before failure?}
}
\description{
This is an internal function used by \code{\link{qr_scan_js_from_corners}}
and \code{\link{qr_scan_cpp}}. It's a wrapper around \code{\link[progress]{progress_bar}}.
It checks for \code{options("dplyr.show_progress")} since that's a handy flag
used by an existing dependency of this package. If I was smarter, I'd probably
put something into \code{\link{options}} on loading \code{quadrangle}.
}
\keyword{internal}
