% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_scan.R
\name{qr_plot}
\alias{qr_plot}
\title{Plot a \code{magick} image with any detected QR codes.}
\usage{
qr_plot(mgk, code_obj)
}
\arguments{
\item{mgk}{The image object that was scanned.}

\item{code_obj}{A list of dataframes, "values" and "points", generated by \code{qr_scan}.}
}
\value{
A ggplot object.
}
\description{
\code{qr_plot} is a helper called inside \code{qr_scan} to visualize images
and any QR codes. Requires \code{ggplot2}.
}
