% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qr_rbind_}
\alias{qr_rbind_}
\title{(Internal) \code{dplyr::bind_rows} replacement}
\usage{
qr_rbind_(df_list, .id = "id")
}
\arguments{
\item{df_list}{A list of dataframes.}

\item{.id}{The name of the prepended ID column.}
}
\description{
This is an internal function. Since \pkg{dplyr} is a heavy dependency, this
handles the expected use cases in this package. It's inferior to \code{dplyr::bind_rows}
in every way, so it should only be used inside this package.
}
\keyword{internal}
