% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_helpers.R
\name{qr_scan_cpp}
\alias{qr_scan_cpp}
\title{Scan an image file or \code{magick} image object for QR codes with the C++ engine.}
\usage{
qr_scan_cpp(image, flop = T, debug = F)
}
\arguments{
\item{image}{A path to a \code{magick}-readable file, e.g. jpg or png, or a \code{magick} object.}

\item{flop}{Logical. Should image be mirrored L-R? Some generators produce QR codes like this.}

\item{debug}{Logical. Should additional metadata about decoded QR patterns be included? e.g. ECC level, version number, etc.}
}
\value{
A list of two dataframes, "values" and "points" describing any found QR codes.
}
\description{
This is a wrapper function to call the QR code scanner in the \code{quirc}
C++ library. You can call this function on file paths, or preprocess files
and call this function on the resulting \code{magick} objects.
}
