% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_scan_js_array}
\alias{qr_scan_js_array}
\title{Scan a pixel array for QR codes with the JS engine.}
\usage{
qr_scan_js_array(arr, engine = NULL)
}
\arguments{
\item{arr}{A 3D array of types raw, double, or integer.}

\item{engine}{An initialized V8 context, see \code{\link{qr_v8_init}}.}
}
\value{
A list with metadata about any identified QR code.
}
\description{
This is the wrapper for the QR code scanner in the \code{jsQR} library. It
converts an array of pixels (usually from \code{\link{image_data}}) into the
appropriate JS object and calls \code{jsQR} to scan it. The first dimension
should length-4 for RGBA values (can be \code{double} 0-1, \code{integer} 0-255, or \code{raw}
\code{0x00-0xff}), and the other dimensions should be width and height.
}
