% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpp_helpers.R
\name{qr_threshold_shortcut_}
\alias{qr_threshold_shortcut_}
\title{(Internal) Image thresholding, with some shortcuts}
\usage{
qr_threshold_shortcut_(mgk, type, threshold, ...)
}
\arguments{
\item{mgk}{A magick image object.}

\item{type}{Type of thresholding, either black or white.}

\item{threshold}{Pixel intensity threshold.}

\item{...}{Additional arguments passed through (\code{channels}).}
}
\description{
This is an internal function used by \code{\link{qr_scan_js_from_corners}}
and \code{\link{qr_scan_cpp}}. It's a wrapper around \code{\link{image_threshold}}
which skips running the function call if the arguments indicate essentially
unchanged images, such as pushing the blackest 0\% of pixels to black, or the
whitest 100\% of pixels to white.
}
\keyword{internal}
