% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_v8_checker_}
\alias{qr_v8_checker_}
\title{Check if the system V8 library is up to date.}
\usage{
qr_v8_checker_()
}
\value{
Logical. Compatibility.
}
\description{
Older versions of V8 included with Ubuntu/Debian do not support 
\code{Uint8ClampedArray} typed arrays, so JS decoding will be skipped. These
can be updated manually.
}
\keyword{internal}
