% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_v8_ls}
\alias{qr_v8_ls}
\title{Identify a JS engine and test it.}
\usage{
qr_v8_ls(env = .GlobalEnv, qr_js_src = qr_js_src_path)
}
\arguments{
\item{env}{Which environment to search for engines? (Usually global.)}

\item{qr_js_src}{Path to the \code{jsQR} library source.}
}
\value{
A prepared \code{\link{v8}} context.
}
\description{
If a \pkg{V8} engine isn't specified when running \code{\link{qr_scan_js_array}}, this
function will be called to search the global environment. If it finds a V8 
context, it will then test if it contains the \code{jsQR} library and return
it. If not, it will initialize a new one with \code{\link{qr_v8_init}} and return it.
}
