% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_qr_scan_array}
\alias{rcpp_qr_scan_array}
\title{Scan a grayscale array using the quirc C++ engine}
\usage{
rcpp_qr_scan_array(grayarray, w, h, debug = FALSE)
}
\arguments{
\item{grayarray}{Vector of grayscale pixel values}

\item{w}{Width of pixel array}

\item{h}{Height of pixel array}

\item{debug}{Return additional info about found QR codes (ECC level, version, etc)}
}
\description{
Should only be called from quadrangle::qr_scan_cpp in R
}
