% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_js.R
\name{set_missing_text_to_latin1_}
\alias{set_missing_text_to_latin1_}
\title{Set missing text from bytes assuming latin1}
\usage{
set_missing_text_to_latin1_(lst)
}
\arguments{
\item{lst}{A list with dataframe element \strong{values}, used in the internal of \code{\link{qr_parse_js_}}.}
}
\value{
A list with dataframe element \strong{values}.
}
\description{
This function is only called by \code{\link{qr_parse_js_}}, to fill missing
text entries from the bytes integer vector.
}
\keyword{internal}
