% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_js.R
\name{qr_parse_js_}
\alias{qr_parse_js_}
\title{Parse multiple QR objects into a single object.}
\usage{
qr_parse_js_(lst)
}
\arguments{
\item{lst}{A list returned by \code{\link{qr_scan_js_from_corners}}, with dataframe elements \strong{chunks} and \strong{location}.}
}
\value{
If not empty, a list with two dataframe elements, \strong{values} and \strong{points}, identifying all QR codes from an image.
}
\description{
This function is usually only called by \code{\link{qr_scan}}, to combine multiple
objects returned by \code{\link{qr_scan_js_from_corners}} into a single dataframe,
analogous to the output from \code{rcpp_qr_scan_array} or \code{\link{qr_scan_cpp}}.
}
\keyword{internal}
