% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{qr_v8_init}
\alias{qr_v8_init}
\title{Initialize a JS engine for QR scanning.}
\usage{
qr_v8_init()
}
\value{
A V8 context that can be used inside a function, or assigned to the global environment.
}
\description{
Initialize a JS engine for QR scanning.
}
